+++ __tests__/server/singular.js
@@ -21,26 +21,37 @@ describe('Server', () => {
   })
 
   describe('GET /:resource', () => {
-    it('should respond with corresponding resource', () =>
-      request(server).get('/user').expect(db.user).expect(200))
+    test('should respond with corresponding resource', () =>
+      request(server)
+        .get('/user')
+        .expect(db.user)
+        .expect(200))
   })
 
   describe('POST /:resource', () => {
-    it('should create resource', () => {
+    test('should create resource', () => {
       const user = { name: 'bar' }
-      return request(server).post('/user').send(user).expect(user).expect(201)
+      return request(server)
+        .post('/user')
+        .send(user)
+        .expect(user)
+        .expect(201)
     })
   })
 
   describe('PUT /:resource', () => {
-    it('should update resource', () => {
+    test('should update resource', () => {
       const user = { name: 'bar' }
-      return request(server).put('/user').send(user).expect(user).expect(200)
+      return request(server)
+        .put('/user')
+        .send(user)
+        .expect(user)
+        .expect(200)
     })
   })
 
   describe('PATCH /:resource', () => {
-    it('should update resource', () =>
+    test('should update resource', () =>
       request(server)
         .patch('/user')
         .send({ name: 'bar' })