+++ __tests__/server/utils.js
@@ -6,7 +6,7 @@ describe('utils', () => {
     const array = [1, 2, 3, 4, 5]
     const perPage = 2
 
-    it('should return first page', () => {
+    test('should return first page', () => {
       assert.deepEqual(utils.getPage(array, 1, perPage), {
         items: [1, 2],
         current: 1,
@@ -16,7 +16,7 @@ describe('utils', () => {
       })
     })
 
-    it('should return second page', () => {
+    test('should return second page', () => {
       assert.deepEqual(utils.getPage(array, 2, perPage), {
         items: [3, 4],
         current: 2,
@@ -27,7 +27,7 @@ describe('utils', () => {
       })
     })
 
-    it('should return third page (last)', () => {
+    test('should return third page (last)', () => {
       assert.deepEqual(utils.getPage(array, 3, perPage), {
         items: [5],
         current: 3,
@@ -37,19 +37,19 @@ describe('utils', () => {
       })
     })
 
-    it('should return an empty array if page is greater than the last page', () => {
+    test('should return an empty array if page is greater than the last page', () => {
       assert.deepEqual(utils.getPage(array, 99, perPage), {
         items: []
       })
     })
 
-    it('should return the array if perPage is greater than the array size', () => {
+    test('should return the array if perPage is greater than the array size', () => {
       assert.deepEqual(utils.getPage(array, 1, 99), {
         items: array
       })
     })
 
-    it('should return an empty array if the array is empty', () => {
+    test('should return an empty array if the array is empty', () => {
       assert.deepEqual(utils.getPage([], 1, 1), {
         items: []
       })