+++ bin/index.js
@@ -1,3 +1,0 @@
-#!/usr/bin/env node
-require('please-upgrade-node')(require('../package.json'))
-require('../lib/cli/bin')