+++ src/server/mixins.js
@@ -1,4 +1,4 @@
-const shortid = require('shortid')
+const nanoid = require('nanoid')
 const pluralize = require('pluralize')
 
 module.exports = {
@@ -48,7 +48,7 @@ function createId(coll) {
     let id = _(coll).maxBy(idProperty)[idProperty]
 
     // Increment integer id or generate string id
-    return _.isFinite(id) ? ++id : shortid.generate()
+    return _.isFinite(id) ? ++id : nanoid(7)
   }
 }
 
@@ -67,7 +67,12 @@ function deepQuery(value, q) {
           return true
         }
       }
-    } else if (value.toString().toLowerCase().indexOf(q) !== -1) {
+    } else if (
+      value
+        .toString()
+        .toLowerCase()
+        .indexOf(q) !== -1
+    ) {
       return true
     }
   }