+++ src/server/router/plural.js
@@ -4,10 +4,12 @@ const pluralize = require('pluralize')
 const write = require('./write')
 const getFullURL = require('./get-full-url')
 const utils = require('../utils')
+const delay = require('./delay')
 
 module.exports = (db, name, opts) => {
   // Create router
   const router = express.Router()
+  router.use(delay)
 
   // Embed function used in GET /name and GET /name/id
   function embed(resource, e) {
@@ -234,7 +236,10 @@ module.exports = (db, name, opts) => {
   function show(req, res, next) {
     const _embed = req.query._embed
     const _expand = req.query._expand
-    const resource = db.get(name).getById(req.params.id).value()
+    const resource = db
+      .get(name)
+      .getById(req.params.id)
+      .value()
 
     if (resource) {
       // Clone resource to avoid making changes to the underlying object
@@ -256,7 +261,10 @@ module.exports = (db, name, opts) => {
 
   // POST /name
   function create(req, res, next) {
-    const resource = db.get(name).insert(req.body).value()
+    const resource = db
+      .get(name)
+      .insert(req.body)
+      .value()
 
     res.setHeader('Access-Control-Expose-Headers', 'Location')
     res.location(`${getFullURL(req)}/${resource.id}`)
@@ -289,14 +297,17 @@ module.exports = (db, name, opts) => {
 
   // DELETE /name/:id
   function destroy(req, res, next) {
-    const resource = db.get(name).removeById(req.params.id).value()
+    const resource = db
+      .get(name)
+      .removeById(req.params.id)
+      .value()
 
     // Remove dependents documents
-    console.log({ opts })
     const removable = db._.getRemovable(db.getState(), opts)
-    console.log(removable)
     removable.forEach(item => {
-      db.get(item.name).removeById(item.id).value()
+      db.get(item.name)
+        .removeById(item.id)
+        .value()
     })
 
     if (resource) {
@@ -308,7 +319,10 @@ module.exports = (db, name, opts) => {
 
   const w = write(db)
 
-  router.route('/').get(list).post(create, w)
+  router
+    .route('/')
+    .get(list)
+    .post(create, w)
 
   router
     .route('/:id')