+++ src/server/router/singular.js
@@ -1,9 +1,11 @@
 const express = require('express')
 const write = require('./write')
 const getFullURL = require('./get-full-url')
+const delay = require('./delay')
 
 module.exports = (db, name) => {
   const router = express.Router()
+  router.use(delay)
 
   function show(req, res, next) {
     res.locals.data = db.get(name).value()
@@ -25,7 +27,9 @@ module.exports = (db, name) => {
     if (req.method === 'PUT') {
       db.set(name, req.body).value()
     } else {
-      db.get(name).assign(req.body).value()
+      db.get(name)
+        .assign(req.body)
+        .value()
     }
 
     res.locals.data = db.get(name).value()
@@ -34,7 +38,12 @@ module.exports = (db, name) => {
 
   const w = write(db)
 
-  router.route('/').get(show).post(create, w).put(update, w).patch(update, w)
+  router
+    .route('/')
+    .get(show)
+    .post(create, w)
+    .put(update, w)
+    .patch(update, w)
 
   return router
 }