+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Change Log
 
+## 0.14.0 - 2018-06-09
+
+* Listen to `localhost` by default, instead of `0.0.0.0`
+
+## 0.13.0 - 2018-05-30
+
+* Bundle all index page assets so that you access it without network connection
+* Drop Node 4 support
+
+## 0.12.2 - 2018-04-26
+
+* Add `_delay` query parameter
+* Upgrade `please-upgrade-node` dependency
+
+## 0.12.1 - 2017-11-02
+
+* Disable logging for static content requests
+* Remove bad `console.log`
+* Update `dependencies`
+* Use [`nanoid`](https://github.com/ai/nanoid)
+
 ## 0.12.0 - 2017-08-02
 
 Re-include `body-parser` in `jsonServer.defaults()`